feat: 관심 채널 순서 변경 API 연동
- 관심 채널 순서 변경 API 호출하는 updateInterestChannel 함수 추가
- handleChannelDragEnd 유틸 함수로 분리하여 관심 채널 순서 변경 처리
